move handler functions into control panel item components
