+++ src/ControlPanel/Tempo.tsx
@@ -1,20 +1,33 @@
+import { useState } from 'react'
 import ControlPanelItem from './ControlPanelItem'
 
 type TempoProps = {
-  handleChange: React.ChangeEventHandler<HTMLInputElement>
+  onChange(bpm: number): void
   defaultValue: number
-  value: string
 }
 export default function Tempo(props: TempoProps) {
+  const [visualBpm, setVisualBpm] = useState(120)
+  // TODO: something about this isn't working right
+  const handleChange: React.ChangeEventHandler<HTMLInputElement> = (event) => {
+    const bpm = Number(event.target.value)
+    if (Number.isNaN(bpm)) {
+      throw new Error(
+        `Could not convert bpm "${event.target.value}" to numeric`
+      )
+    }
+    props.onChange(bpm)
+    setVisualBpm(bpm)
+  }
+
   return (
     <ControlPanelItem>
       <div>
-        <span className="font-serif text-4xl pr-3">{props.value}</span>
+        <span className="font-serif text-4xl pr-3">{visualBpm}</span>
         <span className="font-serif text-xl">BPM</span>
       </div>
       <input
         type="range"
-        onChange={props.handleChange}
+        onChange={handleChange}
         min={20}
         max={300}
         step={0.1}
