+++ src/ControlPanel/TimeSignature.tsx
@@ -1,15 +1,36 @@
+import { TimeSignature as TimeSignatureType } from '../Metronome'
 import ControlPanelItem from './ControlPanelItem'
 
 type TimeSignatureProps = {
-  handleChange: React.ChangeEventHandler<HTMLSelectElement>
+  onChange(signature: TimeSignatureType): void
   beatsPerMeasure: number
   beatUnit: number
 }
 export default function TimeSignature(props: TimeSignatureProps) {
+  const handleChange: React.ChangeEventHandler<HTMLSelectElement> = (event) => {
+    const [beatsPerMeasureStr, beatUnitStr] = event.target.value?.split('/')
+    if (!beatsPerMeasureStr || !beatUnitStr) {
+      throw new Error(`Could not parse time signature "${event.target.value}"`)
+    }
+    const [beatsPerMeasure, beatUnit] = [
+      Number(beatsPerMeasureStr),
+      Number(beatUnitStr),
+    ]
+    if (Number.isNaN(beatsPerMeasure) || Number.isNaN(beatUnit)) {
+      throw new Error(
+        `Could not convert time signature "${event.target.value}" to numeric values`
+      )
+    }
+    props.onChange({
+      beatsPerMeasure,
+      beatUnit,
+    })
+  }
+
   return (
     <ControlPanelItem>
       <select
-        onChange={props.handleChange}
+        onChange={handleChange}
         value={`${props.beatsPerMeasure}/${props.beatUnit}`}
         className="font-serif text-4xl bg-white"
       >
