+++ src/ControlPanel/index.tsx
@@ -16,53 +16,10 @@ export const ControlPanel: React.FC<Props> = ({
   metronome,
   metronomeWriter,
 }) => {
-  const [visualBpm, setVisualBpm] = useState(120)
-  const setUpstreamBpm = useDebouncedCallback(
-    (bpm: number) => {
-      metronomeWriter.setBpm(bpm)
-    },
-    100,
-    { leading: true, trailing: false }
-  )
-  // TODO: something about this isn't working right
-  const handleSetBpm: React.ChangeEventHandler<HTMLInputElement> = (event) => {
-    const bpm = Number(event.target.value)
-    if (Number.isNaN(bpm)) {
-      throw new Error(
-        `Could not convert bpm "${event.target.value}" to numeric`
-      )
-    }
-    setUpstreamBpm(bpm)
-    setVisualBpm(bpm)
-  }
-
-  const handleSetTimeSignature: React.ChangeEventHandler<HTMLSelectElement> = (
-    event
-  ) => {
-    const [beatsPerMeasureStr, beatUnitStr] = event.target.value?.split('/')
-    if (!beatsPerMeasureStr || !beatUnitStr) {
-      throw new Error(`Could not parse time signature "${event.target.value}"`)
-    }
-    const [beatsPerMeasure, beatUnit] = [
-      Number(beatsPerMeasureStr),
-      Number(beatUnitStr),
-    ]
-    if (Number.isNaN(beatsPerMeasure) || Number.isNaN(beatUnit)) {
-      throw new Error(
-        `Could not convert time signature "${event.target.value}" to numeric values`
-      )
-    }
-    metronomeWriter.setTimeSignature({
-      beatsPerMeasure,
-      beatUnit,
-    })
-  }
-
-  const handleChangeMeasureCount: React.ChangeEventHandler<HTMLInputElement> = (
-    event
-  ) => {
-    metronomeWriter.setMeasureCount(Number(event.target.value))
-  }
+  const setUpstreamBpm = useDebouncedCallback(metronomeWriter.setBpm, 100, {
+    leading: true,
+    trailing: false,
+  })
 
   return (
     <div className="flex flex-col">
@@ -72,20 +29,16 @@ export const ControlPanel: React.FC<Props> = ({
           currentMeasure={metronome.currentMeasure}
         />
 
-        <Tempo
-          handleChange={handleSetBpm}
-          defaultValue={metronome.bpm}
-          value={visualBpm.toFixed(1)}
-        />
+        <Tempo onChange={setUpstreamBpm} defaultValue={metronome.bpm} />
 
         <TimeSignature
-          handleChange={handleSetTimeSignature}
+          onChange={metronomeWriter.setTimeSignature}
           beatsPerMeasure={metronome.timeSignature.beatsPerMeasure}
           beatUnit={metronome.timeSignature.beatUnit}
         />
 
         <MeasureCount
-          handleChange={handleChangeMeasureCount}
+          onChange={metronomeWriter.setMeasureCount}
           measureCount={metronome.measureCount}
         />
       </div>
